requirements.txt
cycler==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.0
mlxtend==0.17.2
numpy==1.19.0
pandas==1.0.5
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
threadpoolctl==2.1.0
